propose moving trait implementations next to type declaration
