+++ zingolib/src/wallet/transactions.rs
@@ -14,10 +14,52 @@ impl std::ops::Deref for TransactionRecordMap {
         &self.0
     }
 }
-
 impl std::ops::DerefMut for TransactionRecordMap {
     fn deref_mut(&mut self) -> &mut Self::Target {
         &mut self.0
+    }
+}
+
+impl zcash_client_backend::data_api::InputSource for TransactionRecordMap {
+    type Error = crate::error::ZingoLibError;
+    type AccountId = zcash_primitives::zip32::AccountId;
+    type NoteRef = super::notes::NoteRecordIdentifier;
+
+    fn get_spendable_note(
+        &self,
+        txid: &zcash_primitives::transaction::TxId,
+        protocol: zcash_client_backend::ShieldedProtocol,
+        index: u32,
+    ) -> Result<
+        Option<
+            zcash_client_backend::wallet::ReceivedNote<
+                Self::NoteRef,
+                zcash_client_backend::wallet::Note,
+            >,
+        >,
+        Self::Error,
+    > {
+        let note_record_reference: <Self as zcash_client_backend::data_api::InputSource>::NoteRef =
+            super::notes::NoteRecordIdentifier {
+                txid: *txid,
+                pool: zcash_client_backend::PoolType::Shielded(protocol),
+                index,
+            };
+        Ok(self.get_shielded_received_note_from_identifier(note_record_reference))
+    }
+
+    fn select_spendable_notes(
+        &self,
+        account: Self::AccountId,
+        target_value: zcash_primitives::transaction::components::amount::NonNegativeAmount,
+        sources: &[zcash_client_backend::ShieldedProtocol],
+        anchor_height: zcash_primitives::consensus::BlockHeight,
+        exclude: &[Self::NoteRef],
+    ) -> Result<
+        zcash_client_backend::data_api::SpendableNotes<super::notes::NoteRecordIdentifier>,
+        crate::error::ZingoLibError,
+    > {
+        todo!()
     }
 }
 impl TransactionRecordMap {
